Split arithmetic tests from test_math.py
